@@ -16,15 +16,18 @@
 
 package com.stfalcon.chatkit.dialogs;
 
+import android.graphics.Typeface;
 import android.graphics.drawable.GradientDrawable;
-import android.support.annotation.LayoutRes;
-import android.support.v7.widget.RecyclerView;
 import android.util.TypedValue;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import androidx.annotation.LayoutRes;
+import androidx.annotation.Nullable;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.stfalcon.chatkit.R;
 import com.stfalcon.chatkit.commons.ImageLoader;
@@ -46,10 +49,11 @@ import static android.view.View.VISIBLE;
 /**
  * Adapter for {@link DialogsList}
  */
+@SuppressWarnings("WeakerAccess")
 public class DialogsListAdapter<DIALOG extends IDialog>
         extends RecyclerView.Adapter<DialogsListAdapter.BaseDialogViewHolder> {
 
-    private List<DIALOG> items = new ArrayList<>();
+    protected List<DIALOG> items = new ArrayList<>();
     private int itemLayoutId;
     private Class<? extends BaseDialogViewHolder> holderClass;
     private ImageLoader imageLoader;
@@ -197,18 +201,30 @@ public class DialogsListAdapter<DIALOG extends IDialog>
      */
     public void addItem(DIALOG dialog) {
         items.add(dialog);
-        notifyItemInserted(0);
+        notifyItemInserted(items.size() - 1);
     }
 
     /**
      * Add dialog to dialogs list
      *
      * @param dialog   dialog item
-     * @param position position in dialogs lost
+     * @param position position in dialogs list
      */
     public void addItem(int position, DIALOG dialog) {
         items.add(position, dialog);
         notifyItemInserted(position);
+    }
+
+    /**
+     * Move an item
+     *
+     * @param fromPosition the actual position of the item
+     * @param toPosition   the new position of the item
+     */
+    public void moveItem(int fromPosition, int toPosition) {
+        DIALOG dialog = items.remove(fromPosition);
+        items.add(toPosition, dialog);
+        notifyItemMoved(fromPosition, toPosition);
     }
 
     /**
@@ -244,6 +260,47 @@ public class DialogsListAdapter<DIALOG extends IDialog>
     }
 
     /**
+     * Upsert dialog in dialogs list or add it to then end of dialogs list
+     *
+     * @param item dialog item
+     */
+    public void upsertItem(DIALOG item) {
+        boolean updated = false;
+        for (int i = 0; i < items.size(); i++) {
+            if (items.get(i).getId().equals(item.getId())) {
+                items.set(i, item);
+                notifyItemChanged(i);
+                updated = true;
+                break;
+            }
+        }
+        if (!updated) {
+            addItem(item);
+        }
+    }
+
+    /**
+     * Find an item by its id
+     *
+     * @param id the wanted item's id
+     * @return the found item, or null
+     */
+    @Nullable
+    public DIALOG getItemById(String id) {
+        if (items == null) {
+            items = new ArrayList<>();
+        }
+        for (DIALOG item : items) {
+            if (item.getId() == null && id == null) {
+                return item;
+            } else if (item.getId() != null && item.getId().equals(id)) {
+                return item;
+            }
+        }
+        return null;
+    }
+
+    /**
      * Update last message in dialog and swap item to top of list.
      *
      * @param dialogId Dialog ID
@@ -272,15 +329,12 @@ public class DialogsListAdapter<DIALOG extends IDialog>
      * Sort dialog by last message date
      */
     public void sortByLastMessageDate() {
-        Collections.sort(items, new Comparator<DIALOG>() {
-            @Override
-            public int compare(DIALOG o1, DIALOG o2) {
-                if (o1.getLastMessage().getCreatedAt().after(o2.getLastMessage().getCreatedAt())) {
-                    return -1;
-                } else if (o1.getLastMessage().getCreatedAt().before(o2.getLastMessage().getCreatedAt())) {
-                    return 1;
-                } else return 0;
-            }
+        Collections.sort(items, (o1, o2) -> {
+            if (o1.getLastMessage().getCreatedAt().after(o2.getLastMessage().getCreatedAt())) {
+                return -1;
+            } else if (o1.getLastMessage().getCreatedAt().before(o2.getLastMessage().getCreatedAt())) {
+                return 1;
+            } else return 0;
         });
         notifyDataSetChanged();
     }
@@ -296,19 +350,19 @@ public class DialogsListAdapter<DIALOG extends IDialog>
     }
 
     /**
+     * @return registered image loader
+     */
+    public ImageLoader getImageLoader() {
+        return imageLoader;
+    }
+
+    /**
      * Register a callback to be invoked when image need to load.
      *
      * @param imageLoader image loading method
      */
     public void setImageLoader(ImageLoader imageLoader) {
         this.imageLoader = imageLoader;
-    }
-
-    /**
-     * @return registered image loader
-     */
-    public ImageLoader getImageLoader() {
-        return imageLoader;
     }
 
     /**
@@ -387,9 +441,16 @@ public class DialogsListAdapter<DIALOG extends IDialog>
         this.dialogStyle = dialogStyle;
     }
 
+    /**
+     * @return the position of a dialog in the dialogs list.
+     */
+    public int getDialogPosition(DIALOG dialog) {
+        return this.items.indexOf(dialog);
+    }
+
     /*
-    * LISTENERS
-    * */
+     * LISTENERS
+     * */
     public interface OnDialogClickListener<DIALOG extends IDialog> {
         void onDialogClick(DIALOG dialog);
     }
@@ -407,8 +468,8 @@ public class DialogsListAdapter<DIALOG extends IDialog>
     }
 
     /*
-    * HOLDERS
-    * */
+     * HOLDERS
+     * */
     public abstract static class BaseDialogViewHolder<DIALOG extends IDialog>
             extends ViewHolder<DIALOG> {
 
@@ -427,19 +488,19 @@ public class DialogsListAdapter<DIALOG extends IDialog>
             this.imageLoader = imageLoader;
         }
 
-        void setOnDialogClickListener(OnDialogClickListener<DIALOG> onDialogClickListener) {
+        protected void setOnDialogClickListener(OnDialogClickListener<DIALOG> onDialogClickListener) {
             this.onDialogClickListener = onDialogClickListener;
         }
 
-        void setOnDialogViewClickListener(OnDialogViewClickListener<DIALOG> onDialogViewClickListener) {
+        protected void setOnDialogViewClickListener(OnDialogViewClickListener<DIALOG> onDialogViewClickListener) {
             this.onDialogViewClickListener = onDialogViewClickListener;
         }
 
-        void setOnLongItemClickListener(OnDialogLongClickListener<DIALOG> onLongItemClickListener) {
+        protected void setOnLongItemClickListener(OnDialogLongClickListener<DIALOG> onLongItemClickListener) {
             this.onLongItemClickListener = onLongItemClickListener;
         }
 
-        void setOnDialogViewLongClickListener(OnDialogViewLongClickListener<DIALOG> onDialogViewLongClickListener) {
+        protected void setOnDialogViewLongClickListener(OnDialogViewLongClickListener<DIALOG> onDialogViewLongClickListener) {
             this.onDialogViewLongClickListener = onDialogViewLongClickListener;
         }
 
@@ -463,15 +524,15 @@ public class DialogsListAdapter<DIALOG extends IDialog>
 
         public DialogViewHolder(View itemView) {
             super(itemView);
-            root = (ViewGroup) itemView.findViewById(R.id.dialogRootLayout);
-            container = (ViewGroup) itemView.findViewById(R.id.dialogContainer);
-            tvName = (TextView) itemView.findViewById(R.id.dialogName);
-            tvDate = (TextView) itemView.findViewById(R.id.dialogDate);
-            tvLastMessage = (TextView) itemView.findViewById(R.id.dialogLastMessage);
-            tvBubble = (TextView) itemView.findViewById(R.id.dialogUnreadBubble);
-            ivLastMessageUser = (ImageView) itemView.findViewById(R.id.dialogLastMessageUserAvatar);
-            ivAvatar = (ImageView) itemView.findViewById(R.id.dialogAvatar);
-            dividerContainer = (ViewGroup) itemView.findViewById(R.id.dialogDividerContainer);
+            root = itemView.findViewById(R.id.dialogRootLayout);
+            container = itemView.findViewById(R.id.dialogContainer);
+            tvName = itemView.findViewById(R.id.dialogName);
+            tvDate = itemView.findViewById(R.id.dialogDate);
+            tvLastMessage = itemView.findViewById(R.id.dialogLastMessage);
+            tvBubble = itemView.findViewById(R.id.dialogUnreadBubble);
+            ivLastMessageUser = itemView.findViewById(R.id.dialogLastMessageUserAvatar);
+            ivAvatar = itemView.findViewById(R.id.dialogAvatar);
+            dividerContainer = itemView.findViewById(R.id.dialogDividerContainer);
             divider = itemView.findViewById(R.id.dialogDivider);
 
         }
@@ -530,17 +591,17 @@ public class DialogsListAdapter<DIALOG extends IDialog>
 
                 if (tvName != null) {
                     tvName.setTextColor(dialogStyle.getDialogTitleTextColor());
-                    tvName.setTypeface(tvName.getTypeface(), dialogStyle.getDialogTitleTextStyle());
+                    tvName.setTypeface(Typeface.DEFAULT, dialogStyle.getDialogTitleTextStyle());
                 }
 
                 if (tvDate != null) {
                     tvDate.setTextColor(dialogStyle.getDialogDateColor());
-                    tvDate.setTypeface(tvDate.getTypeface(), dialogStyle.getDialogDateStyle());
+                    tvDate.setTypeface(Typeface.DEFAULT, dialogStyle.getDialogDateStyle());
                 }
 
                 if (tvLastMessage != null) {
                     tvLastMessage.setTextColor(dialogStyle.getDialogMessageTextColor());
-                    tvLastMessage.setTypeface(tvLastMessage.getTypeface(), dialogStyle.getDialogMessageTextStyle());
+                    tvLastMessage.setTypeface(Typeface.DEFAULT, dialogStyle.getDialogMessageTextStyle());
                 }
             }
         }
@@ -553,17 +614,17 @@ public class DialogsListAdapter<DIALOG extends IDialog>
 
                 if (tvName != null) {
                     tvName.setTextColor(dialogStyle.getDialogUnreadTitleTextColor());
-                    tvName.setTypeface(tvName.getTypeface(), dialogStyle.getDialogUnreadTitleTextStyle());
+                    tvName.setTypeface(Typeface.DEFAULT, dialogStyle.getDialogUnreadTitleTextStyle());
                 }
 
                 if (tvDate != null) {
                     tvDate.setTextColor(dialogStyle.getDialogUnreadDateColor());
-                    tvDate.setTypeface(tvDate.getTypeface(), dialogStyle.getDialogUnreadDateStyle());
+                    tvDate.setTypeface(Typeface.DEFAULT, dialogStyle.getDialogUnreadDateStyle());
                 }
 
                 if (tvLastMessage != null) {
                     tvLastMessage.setTextColor(dialogStyle.getDialogUnreadMessageTextColor());
-                    tvLastMessage.setTypeface(tvLastMessage.getTypeface(), dialogStyle.getDialogUnreadMessageTextStyle());
+                    tvLastMessage.setTypeface(Typeface.DEFAULT, dialogStyle.getDialogUnreadMessageTextStyle());
                 }
             }
         }
@@ -582,56 +643,60 @@ public class DialogsListAdapter<DIALOG extends IDialog>
 
             //Set Date
             String formattedDate = null;
-            Date lastMessageDate = dialog.getLastMessage().getCreatedAt();
-            if (datesFormatter != null) formattedDate = datesFormatter.format(lastMessageDate);
-            tvDate.setText(formattedDate == null
-                    ? getDateString(lastMessageDate)
-                    : formattedDate);
+
+            if (dialog.getLastMessage() != null) {
+                Date lastMessageDate = dialog.getLastMessage().getCreatedAt();
+                if (datesFormatter != null) formattedDate = datesFormatter.format(lastMessageDate);
+                tvDate.setText(formattedDate == null
+                        ? getDateString(lastMessageDate)
+                        : formattedDate);
+            } else {
+                tvDate.setText(null);
+            }
 
             //Set Dialog avatar
             if (imageLoader != null) {
-                imageLoader.loadImage(ivAvatar, dialog.getDialogPhoto());
+                imageLoader.loadImage(ivAvatar, dialog.getDialogPhoto(), null);
             }
 
-            //Set Last message user avatar
-            if (imageLoader != null) {
-                imageLoader.loadImage(ivLastMessageUser, dialog.getLastMessage().getUser().getAvatar());
+            //Set Last message user avatar with check if there is last message
+            if (imageLoader != null && dialog.getLastMessage() != null) {
+                imageLoader.loadImage(ivLastMessageUser, dialog.getLastMessage().getUser().getAvatar(), null);
             }
             ivLastMessageUser.setVisibility(dialogStyle.isDialogMessageAvatarEnabled()
-                    && dialog.getUsers().size() > 1 ? VISIBLE : GONE);
+                    && dialog.getUsers().size() > 1
+                    && dialog.getLastMessage() != null ? VISIBLE : GONE);
 
             //Set Last message text
-            tvLastMessage.setText(dialog.getLastMessage().getText());
+            if (dialog.getLastMessage() != null) {
+                tvLastMessage.setText(dialog.getLastMessage().getText());
+            } else {
+                tvLastMessage.setText(null);
+            }
 
             //Set Unread message count bubble
             tvBubble.setText(String.valueOf(dialog.getUnreadCount()));
             tvBubble.setVisibility(dialogStyle.isDialogUnreadBubbleEnabled() &&
                     dialog.getUnreadCount() > 0 ? VISIBLE : GONE);
 
-            container.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View view) {
-                    if (onDialogClickListener != null) {
-                        onDialogClickListener.onDialogClick(dialog);
-                    }
-                    if (onDialogViewClickListener != null) {
-                        onDialogViewClickListener.onDialogViewClick(view, dialog);
-                    }
+            container.setOnClickListener(view -> {
+                if (onDialogClickListener != null) {
+                    onDialogClickListener.onDialogClick(dialog);
+                }
+                if (onDialogViewClickListener != null) {
+                    onDialogViewClickListener.onDialogViewClick(view, dialog);
                 }
             });
 
 
-            container.setOnLongClickListener(new View.OnLongClickListener() {
-                @Override
-                public boolean onLongClick(View view) {
-                    if (onLongItemClickListener != null) {
-                        onLongItemClickListener.onDialogLongClick(dialog);
-                    }
-                    if (onDialogViewLongClickListener != null) {
-                        onDialogViewLongClickListener.onDialogViewLongClick(view, dialog);
-                    }
-                    return onLongItemClickListener != null || onDialogViewLongClickListener != null;
+            container.setOnLongClickListener(view -> {
+                if (onLongItemClickListener != null) {
+                    onLongItemClickListener.onDialogLongClick(dialog);
                 }
+                if (onDialogViewLongClickListener != null) {
+                    onDialogViewLongClickListener.onDialogViewLongClick(view, dialog);
+                }
+                return onLongItemClickListener != null || onDialogViewLongClickListener != null;
             });
         }
 
