@@ -1,9 +1,10 @@
 package com.stfalcon.chatkit.sample.features.main.adapter;
 
 import android.content.Context;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentStatePagerAdapter;
+
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentStatePagerAdapter;
 
 import com.stfalcon.chatkit.sample.R;
 
@@ -18,10 +19,10 @@ public class MainActivityPagerAdapter extends FragmentStatePagerAdapter {
     public static final int ID_CUSTOM_VIEW_HOLDER = 3;
     public static final int ID_CUSTOM_CONTENT = 4;
 
-    private Context context;
+    private final Context context;
 
     public MainActivityPagerAdapter(Context context, FragmentManager fm) {
-        super(fm);
+        super(fm, BEHAVIOR_RESUME_ONLY_CURRENT_FRAGMENT);
         this.context = context;
     }
 
