@@ -19,8 +19,6 @@ package com.stfalcon.chatkit.messages;
 import android.content.Context;
 import android.graphics.drawable.Drawable;
 import android.os.Build;
-import android.support.v4.view.ViewCompat;
-import android.support.v4.widget.Space;
 import android.text.Editable;
 import android.text.TextWatcher;
 import android.util.AttributeSet;
@@ -29,7 +27,10 @@ import android.view.View;
 import android.widget.EditText;
 import android.widget.ImageButton;
 import android.widget.RelativeLayout;
+import android.widget.Space;
 import android.widget.TextView;
+
+import androidx.core.view.ViewCompat;
 
 import com.stfalcon.chatkit.R;
 
@@ -38,8 +39,9 @@ import java.lang.reflect.Field;
 /**
  * Component for input outcoming messages
  */
+@SuppressWarnings({"WeakerAccess", "unused"})
 public class MessageInput extends RelativeLayout
-        implements View.OnClickListener, TextWatcher {
+        implements View.OnClickListener, TextWatcher, View.OnFocusChangeListener {
 
     protected EditText messageInput;
     protected ImageButton messageSendButton;
@@ -49,6 +51,19 @@ public class MessageInput extends RelativeLayout
     private CharSequence input;
     private InputListener inputListener;
     private AttachmentsListener attachmentsListener;
+    private boolean isTyping;
+    private TypingListener typingListener;
+    private int delayTypingStatusMillis;
+    private Runnable typingTimerRunnable = new Runnable() {
+        @Override
+        public void run() {
+            if (isTyping) {
+                isTyping = false;
+                if (typingListener != null) typingListener.onStopTyping();
+            }
+        }
+    };
+    private boolean lastFocus;
 
     public MessageInput(Context context) {
         super(context);
@@ -109,6 +124,8 @@ public class MessageInput extends RelativeLayout
             if (isSubmitted) {
                 messageInput.setText("");
             }
+            removeCallbacks(typingTimerRunnable);
+            post(typingTimerRunnable);
         } else if (id == R.id.attachmentButton) {
             onAddAttachments();
         }
@@ -122,6 +139,14 @@ public class MessageInput extends RelativeLayout
     public void onTextChanged(CharSequence s, int start, int count, int after) {
         input = s;
         messageSendButton.setEnabled(input.length() > 0);
+        if (s.length() > 0) {
+            if (!isTyping) {
+                isTyping = true;
+                if (typingListener != null) typingListener.onStartTyping();
+            }
+            removeCallbacks(typingTimerRunnable);
+            postDelayed(typingTimerRunnable, delayTypingStatusMillis);
+        }
     }
 
     /**
@@ -130,7 +155,7 @@ public class MessageInput extends RelativeLayout
      */
     @Override
     public void beforeTextChanged(CharSequence charSequence, int i, int i1, int i2) {
-
+        //do nothing
     }
 
     /**
@@ -138,7 +163,15 @@ public class MessageInput extends RelativeLayout
      */
     @Override
     public void afterTextChanged(Editable editable) {
+        //do nothing
+    }
 
+    @Override
+    public void onFocusChange(View v, boolean hasFocus) {
+        if (lastFocus && !hasFocus && typingListener != null) {
+            typingListener.onStopTyping();
+        }
+        lastFocus = hasFocus;
     }
 
     private boolean onSubmit() {
@@ -188,21 +221,23 @@ public class MessageInput extends RelativeLayout
                     style.getInputDefaultPaddingBottom()
             );
         }
+        this.delayTypingStatusMillis = style.getDelayTypingStatus();
     }
 
     private void init(Context context) {
         inflate(context, R.layout.view_message_input, this);
 
-        messageInput = (EditText) findViewById(R.id.messageInput);
-        messageSendButton = (ImageButton) findViewById(R.id.messageSendButton);
-        attachmentButton = (ImageButton) findViewById(R.id.attachmentButton);
-        sendButtonSpace = (Space) findViewById(R.id.sendButtonSpace);
-        attachmentButtonSpace = (Space) findViewById(R.id.attachmentButtonSpace);
+        messageInput = findViewById(R.id.messageInput);
+        messageSendButton = findViewById(R.id.messageSendButton);
+        attachmentButton = findViewById(R.id.attachmentButton);
+        sendButtonSpace = findViewById(R.id.sendButtonSpace);
+        attachmentButtonSpace = findViewById(R.id.attachmentButtonSpace);
 
         messageSendButton.setOnClickListener(this);
         attachmentButton.setOnClickListener(this);
         messageInput.addTextChangedListener(this);
         messageInput.setText("");
+        messageInput.setOnFocusChangeListener(this);
     }
 
     private void setCursor(Drawable drawable) {
@@ -230,6 +265,10 @@ public class MessageInput extends RelativeLayout
         }
     }
 
+    public void setTypingListener(TypingListener typingListener) {
+        this.typingListener = typingListener;
+    }
+
     /**
      * Interface definition for a callback to be invoked when user pressed 'submit' button
      */
@@ -253,5 +292,22 @@ public class MessageInput extends RelativeLayout
          * Fires when user presses 'add' button.
          */
         void onAddAttachments();
+    }
+
+    /**
+     * Interface definition for a callback to be invoked when user typing
+     */
+    public interface TypingListener {
+
+        /**
+         * Fires when user presses start typing
+         */
+        void onStartTyping();
+
+        /**
+         * Fires when user presses stop typing
+         */
+        void onStopTyping();
+
     }
 }
