@@ -11,8 +11,8 @@ public class CustomIncomingTextMessageViewHolder
 
     private View onlineIndicator;
 
-    public CustomIncomingTextMessageViewHolder(View itemView) {
-        super(itemView);
+    public CustomIncomingTextMessageViewHolder(View itemView, Object payload) {
+        super(itemView, payload);
         onlineIndicator = itemView.findViewById(R.id.onlineIndicator);
     }
 
@@ -26,5 +26,21 @@ public class CustomIncomingTextMessageViewHolder
         } else {
             onlineIndicator.setBackgroundResource(R.drawable.shape_bubble_offline);
         }
+
+        //We can set click listener on view from payload
+        final Payload payload = (Payload) this.payload;
+        userAvatar.setOnClickListener(view -> {
+            if (payload != null && payload.avatarClickListener != null) {
+                payload.avatarClickListener.onAvatarClick();
+            }
+        });
+    }
+
+    public static class Payload {
+        public OnAvatarClickListener avatarClickListener;
+    }
+
+    public interface OnAvatarClickListener {
+        void onAvatarClick();
     }
 }
