@@ -1,5 +1,6 @@
 package com.stfalcon.chatkit.sample.features.demo.custom.holder.holders.messages;
 
+import android.util.Pair;
 import android.view.View;
 
 import com.stfalcon.chatkit.messages.MessageHolders;
@@ -11,8 +12,8 @@ import com.stfalcon.chatkit.sample.common.data.model.Message;
 public class CustomOutcomingImageMessageViewHolder
         extends MessageHolders.OutcomingImageMessageViewHolder<Message> {
 
-    public CustomOutcomingImageMessageViewHolder(View itemView) {
-        super(itemView);
+    public CustomOutcomingImageMessageViewHolder(View itemView, Object payload) {
+        super(itemView, payload);
     }
 
     @Override
@@ -20,5 +21,12 @@ public class CustomOutcomingImageMessageViewHolder
         super.onBind(message);
 
         time.setText(message.getStatus() + " " + time.getText());
+    }
+
+    //Override this method to have ability to pass custom data in ImageLoader for loading image(not avatar).
+    @Override
+    protected Object getPayloadForImageLoader(Message message) {
+        //For example you can pass size of placeholder before loading
+        return new Pair<>(100, 100);
     }
 }
