@@ -12,11 +12,7 @@ import com.stfalcon.chatkit.sample.common.data.model.Dialog;
 import com.stfalcon.chatkit.sample.common.data.model.Message;
 import com.stfalcon.chatkit.sample.features.demo.DemoDialogsActivity;
 
-import java.util.ArrayList;
-
 public class DefaultDialogsActivity extends DemoDialogsActivity {
-
-    private ArrayList<Dialog> dialogs;
 
     public static void open(Context context) {
         context.startActivity(new Intent(context, DefaultDialogsActivity.class));
@@ -29,7 +25,7 @@ public class DefaultDialogsActivity extends DemoDialogsActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_default_dialogs);
 
-        dialogsList = (DialogsList) findViewById(R.id.dialogsList);
+        dialogsList = findViewById(R.id.dialogsList);
         initAdapter();
     }
 
