@@ -3,6 +3,7 @@ package com.stfalcon.chatkit.sample.features.demo.def;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
+import android.util.Log;
 
 import com.stfalcon.chatkit.messages.MessageInput;
 import com.stfalcon.chatkit.messages.MessagesList;
@@ -10,10 +11,12 @@ import com.stfalcon.chatkit.messages.MessagesListAdapter;
 import com.stfalcon.chatkit.sample.R;
 import com.stfalcon.chatkit.sample.common.data.fixtures.MessagesFixtures;
 import com.stfalcon.chatkit.sample.features.demo.DemoMessagesActivity;
+import com.stfalcon.chatkit.sample.utils.AppUtils;
 
 public class DefaultMessagesActivity extends DemoMessagesActivity
         implements MessageInput.InputListener,
-        MessageInput.AttachmentsListener {
+        MessageInput.AttachmentsListener,
+        MessageInput.TypingListener {
 
     public static void open(Context context) {
         context.startActivity(new Intent(context, DefaultMessagesActivity.class));
@@ -26,11 +29,13 @@ public class DefaultMessagesActivity extends DemoMessagesActivity
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_default_messages);
 
-        this.messagesList = (MessagesList) findViewById(R.id.messagesList);
+        this.messagesList = findViewById(R.id.messagesList);
         initAdapter();
 
-        MessageInput input = (MessageInput) findViewById(R.id.input);
+        MessageInput input = findViewById(R.id.input);
         input.setInputListener(this);
+        input.setTypingListener(this);
+        input.setAttachmentsListener(this);
     }
 
     @Override
@@ -50,6 +55,20 @@ public class DefaultMessagesActivity extends DemoMessagesActivity
         super.messagesAdapter = new MessagesListAdapter<>(super.senderId, super.imageLoader);
         super.messagesAdapter.enableSelectionMode(this);
         super.messagesAdapter.setLoadMoreListener(this);
+        super.messagesAdapter.registerViewClickListener(R.id.messageUserAvatar,
+                (view, message) -> AppUtils.showToast(DefaultMessagesActivity.this,
+                        message.getUser().getName() + " avatar click",
+                        false));
         this.messagesList.setAdapter(super.messagesAdapter);
+    }
+
+    @Override
+    public void onStartTyping() {
+        Log.v("Typing listener", getString(R.string.start_typing_status));
+    }
+
+    @Override
+    public void onStopTyping() {
+        Log.v("Typing listener", getString(R.string.stop_typing_status));
     }
 }
